progress bar section error state
