+++ src/App.tsx
@@ -19,8 +19,8 @@ import DashboardApplicant from "./views/portal/applicant/DashboardApplicant"
 import useAuth from "./hooks/useAuth"
 import Apply from "./views/apply"
 import { Toaster } from "react-hot-toast"
-import UserSection from "./views/apply/userSection"
-import DemographicsSection from "./views/apply/demographicsSection"
+import UserSection from "./views/apply/0 - User"
+import DemographicsSection from "./views/apply/1 - Demographics"
 import ShortResponseSection from "./views/apply/ShortResponseSection"
 import LogisticsSection from "./views/apply/LogisticsSection"
 import SocialsSection from "./views/apply/SocialsSection"
